+++ include/libmfmidi/samhandlers.hpp
@@ -1,6 +1,19 @@
-/// \file samhandlers.cpp
-/// \author Creepercdn (creepercdn@outlook.com)
-/// \brief Useful SAM Handlers
+/*
+ * This file is a part of libmfmidi.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
 
 #pragma once
 
@@ -32,14 +45,14 @@ namespace libmfmidi {
                  << endl;
         }
 
-        void on_starttrack(uint32_t trk) override
+        void on_starttrack(uint16_t trk) override
         {
             ticktime = 0;
             *stm << "Track " << trk << ":" << '\n'
                  << "Tick Time\tMessage Type\tMessage" << std::endl;
         }
 
-        void on_endtrack(uint32_t  /*trk*/) override
+        void on_endtrack(uint16_t  /*trk*/) override
         {
             *stm << std::endl;
         }
@@ -69,12 +82,12 @@ namespace libmfmidi {
             _file->resize(ntrk);
         }
 
-        void on_starttrack(uint32_t trk) override
+        void on_starttrack(uint16_t trk) override
         {
             curtrk = trk;
         }
 
-        void on_endtrack(uint32_t /*trk*/) override
+        void on_endtrack(uint16_t /*trk*/) override
         {
             curtrk = 0xFFFFFFFF;
         }
